requirements.txt
anyio>=4.11.0
google-api-core>=2.26.0
google-api-python-client>=2.184.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
itsdangerous>=2.2.0
matplotlib>=3.10.6
numpy>=2.3.3
oauth2client>=4.1.3
openpyxl>=3.1.5
pandas>=2.3.3
pendulum>=3.1.0
pillow>=11.3.0
pydantic>=2.11.10
pydantic-core>=2.33.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.1.1
qrcode>=8.2
ruff>=0.13.0
uvicorn>=0.37.0
protobuf>=6.33.0